ind_automation/requirements.txt
fastapi
uvicorn
python-docx-template
pydantic
python-dotenv
requests